+++ src/daemon/jdwp.cpp
@@ -147,7 +147,7 @@ void HdcJdwp::ReadStream(uv_stream_t *pipe, ssize_t nread, const uv_buf_t *buf)
             // pid isDebug pkgName/processName
             struct JsMsgHeader *jsMsg = reinterpret_cast<struct JsMsgHeader *>(p);
             if (jsMsg->msgLen == static_cast<uint32_t>(nread)) {
-                pid = jsMsg->pid;
+                pid = static_cast<int32_t>(jsMsg->pid);
                 string pkgName = string((char *)p + sizeof(JsMsgHeader), jsMsg->msgLen - sizeof(JsMsgHeader));
                 ctxJdwp->pkgName = pkgName;
                 ctxJdwp->isDebug = jsMsg->isDebug;
@@ -180,7 +180,7 @@ void HdcJdwp::ReadStream(uv_stream_t *pipe, ssize_t nread, const uv_buf_t *buf)
     }
     (void)memset_s(ctxJdwp->buf, sizeof(ctxJdwp->buf), 0, sizeof(ctxJdwp->buf));
     if (!ret) {
-        WRITE_LOG(LOG_INFO, "ReadStream proc:%d err, free it.", ctxJdwp->pid);
+        WRITE_LOG(LOG_INFO, "ReadStream proc:%u err, free it.", ctxJdwp->pid);
         thisClass->freeContextMutex.lock();
         thisClass->FreeContext(ctxJdwp);
         thisClass->freeContextMutex.unlock();
